+++ backend/database/migrations/2019_09_07_124230_create_announcement_targets_table.php
@@ -13,8 +13,8 @@ class CreateAnnouncementTargetsTable extends Migration {
 	public function up() {
 		Schema::create('announcement_targets', function (Blueprint $table) {
 			$table->bigIncrements('id');
-			$table->unsignedInteger('announcement_id');
-			$table->unsignedInteger('user_id');
+			$table->unsignedBigInteger('announcement_id');
+			$table->unsignedBigInteger('user_id');
 			$table->boolean('is_read')->default(false);
 			$table->timestamps();
