+++ backend/database/migrations/2019_09_13_155204_create_document_targets_table.php
@@ -13,8 +13,8 @@ class CreateDocumentTargetsTable extends Migration {
 	public function up() {
 		Schema::create('document_targets', function (Blueprint $table) {
 			$table->bigIncrements('id');
-			$table->unsignedInteger('document_id');
-			$table->unsignedInteger('user_id');
+			$table->unsignedBigInteger('document_id');
+			$table->unsignedBigInteger('user_id');
 			$table->timestamps();
 
 			$table->foreign('document_id')->references('id')->on('documents')->onDelete('cascade');
